Add podcast_screen_podcast_details_link_tapped event
